Add files via upload
This R script investigates independence between qualitative variables. 
To compare independence between two categorical variables the dominant tests are the chi-squared and the Fisher's exact test. The former can be used only if some further assumptions hold (regarding the number of cells with expected frequencies less than 5).

This script takes two sets of variables: 
one acting as the outcomes and
one acting as the risk factors that we want to associate with the outcomes
It first checks whether the chi-squared test is appropriate. If yes, it is applied, otherwise Fisher's exact test is used.
The output is a list with:
(1) A table with the p-value and the test conducted for each combination between an outcome and a risk factor.
(2) A merged contingency table for each outcome with the risk factors.

The user needs to provide:
1. The dataset
2. A vector of the outcome
3. A vector of the risk factors